+++ CHANGELOG.md
@@ -1,3 +1,12 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+Initial release
+- Select a server to test against (eg: example.com)
+- Configure interval for testing
+- Configure how many failed tests before reporting a connection loss
+- Set latency thresholds for different connection qualities
+- Get notified of connection quality changes in real-time
+- Pause and resume monitoring
+- Query connection history
+- Query current connection status
+- Accommodates occasional network hiccups